requirements.txt
streamlit
pandas
openpyxl
reportlab
Pillow